fix: correct type mismatches in MemorySwap, System, and SystemUser
- Change MemorySwap.free from float to int
- Change System.boot_time and uptime from float to int
- Change SystemUser.started from float to int

These fields are uint64/int in Go but were incorrectly typed as float in Python.

+++ systembridgeconnector/models/modules/memory.py
@@ -14,7 +14,7 @@ class MemorySwap:
 
     total: int | None = None
     used: int | None = None
-    free: float | None = None
+    free: int | None = None
     percent: float | None = None
     sin: int | None = None
     sout: int | None = None
